refactor(ConsoleInputUtil): 클래스명 변경(-> ConsoleInputView)
