+++ src/main/java/study/lotto/step2/presentation/LottoResultsStatistic.java
@@ -1,4 +1,4 @@
-package study.lotto.step2.presentation.output;
+package study.lotto.step2.presentation;
 
 import study.lotto.step2.domain.LottoResult;
 import study.lotto.step2.domain.LottoResults;
